feat: Add lastProjectId to HttpReplicator handler
Update the HttpReplicator handler in the replicator.ts file to include the lastProjectId parameter. This allows for synchronization of scenes from the strips page. The lastProjectId is now included in the fetch URL when calling the endpointPullName. This change enhances the functionality of the replicator and improves the synchronization process.