Feat: Refactor HttpReplicator handler to include lastParagraphInProject
Refactor the HttpReplicator handler in the replicator.ts file to include the lastParagraphInProject parameter. This allows for synchronization of paragraphs from the strips page. The lastParagraphInProject is now included in the fetch URL when calling the endpointPullName. This change enhances the functionality of the replicator and improves the synchronization process.